@@ -1,9 +1,8 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'com.google.gms.google-services'
-apply plugin: "androidx.navigation.safeargs.kotlin"
+apply plugin: "androidx.navigation.safeargs"
 apply plugin: 'com.google.firebase.crashlytics'
 
 kapt {
@@ -11,13 +10,11 @@ kapt {
 }
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
-
+    compileSdk 34
     defaultConfig {
         applicationId "com.task.task"
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 33
         versionCode 20
         versionName "1.0.16"
         multiDexEnabled true
@@ -26,6 +23,9 @@ android {
     }
     lintOptions {
         error("VisibleForTesting")
+    }
+    buildFeatures {
+        viewBinding true
     }
 
     buildTypes {
@@ -51,21 +51,21 @@ android {
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.2'
+    implementation 'androidx.core:core-ktx:1.12.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     implementation 'com.android.support:multidex:1.0.3'
 
     //Material Design
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'com.google.android.material:material:1.9.0'
 
     //Moshi Json parser
-    def moshi_json_parser = '1.11.0'
+    def moshi_json_parser = '1.13.0'
     implementation "com.squareup.moshi:moshi:$moshi_json_parser"
     implementation "com.squareup.moshi:moshi-kotlin:$moshi_json_parser"
 
@@ -76,18 +76,18 @@ dependencies {
 
 
     //Navigation component dependencies
-    def nav_version = "2.3.0"
+    def nav_version = "2.7.3"
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
     implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
-    implementation "androidx.fragment:fragment-ktx:1.3.4"
+    implementation "androidx.fragment:fragment-ktx:1.6.1"
 
     //MvRx dependencies
     def MvRx_version = "1.5.1"
     implementation "com.airbnb.android:mvrx:$MvRx_version"
 
     //Dagger dependencies
-    def dagger_version = "2.27"
+    def dagger_version = "2.43.2"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android-support:$dagger_version"
     annotationProcessor "com.google.dagger:dagger-android-processor:$dagger_version"
@@ -99,8 +99,10 @@ dependencies {
     testImplementation "io.mockk:mockk:1.10.3"
 
     //Android Arch. Components Test
-    def arch_core_test_version = "2.1.0"
+    def arch_core_test_version = "2.2.0"
     testImplementation "androidx.arch.core:core-testing:$arch_core_test_version"
+
+    implementation 'androidx.work:work-runtime:2.9.0-beta01'
 
     implementation project(':Domain')
     implementation project(':LocalDataStore')
