@@ -12,14 +12,11 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.snackbar.Snackbar
 import com.task.task.R
+import com.task.task.databinding.FragmentMoviesListingBinding
 import com.task.task.presentation_module.local_esims.MoviesListViewModel
 import com.task.task.presentation_module.local_esims.models.moviesListEvents
 import com.task.task.ui_module.local_esims_listing.adapter.MoviesListingAdapter
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_movie_details.parentCl
-import kotlinx.android.synthetic.main.fragment_movies_listing.emptyStateView
-import kotlinx.android.synthetic.main.fragment_movies_listing.localeSimsRv
-import kotlinx.android.synthetic.main.fragment_movies_listing.progressBar
 import javax.inject.Inject
 
 class MoviesListingFragment : DaggerFragment(), MoviesListingAdapter.Action {
@@ -30,15 +27,18 @@ class MoviesListingFragment : DaggerFragment(), MoviesListingAdapter.Action {
     private val viewModel: MoviesListViewModel by viewModels { localeSimsViewModelFactory }
 
     private val adapter = MoviesListingAdapter(this)
+
+    private lateinit var binding: FragmentMoviesListingBinding
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_movies_listing, container, false)
+        binding = FragmentMoviesListingBinding.inflate(layoutInflater)
+        return binding.root
     }
 
     private fun bindViews() {
-        localeSimsRv.adapter = adapter
+        binding.localeSimsRv.adapter = adapter
     }
 
     private fun pullData() {
@@ -48,24 +48,24 @@ class MoviesListingFragment : DaggerFragment(), MoviesListingAdapter.Action {
                 when (event) {
                     is moviesListEvents.ErrorState -> {
                         Snackbar.make(
-                            parentCl,
+                            binding.paretCl,
                             "${getString(R.string.something_went_wrong)}${event.err}",
                             Snackbar.LENGTH_SHORT
                         ).show()
                     }
 
                     is moviesListEvents.LoadingState -> {
-                        progressBar.visibility = VISIBLE
+                        binding.progressBar.visibility = VISIBLE
                     }
 
                     is moviesListEvents.RetrievedMoviesListSuccessfully -> {
-                        progressBar.visibility = GONE
+                        binding.progressBar.visibility = GONE
                         with(event.moviesListing.results) {
                             if (isEmpty()) {
-                                emptyStateView.visibility = VISIBLE
+                                binding.emptyStateView.visibility = VISIBLE
                             } else {
                                 adapter.submitList(event.moviesListing.results)
-                                emptyStateView.visibility = GONE
+                                binding.emptyStateView.visibility = GONE
                             }
                         }
                     }
