@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.task.task.R
 import com.task.domain.domain_module.country_listing.models.ResultDomain
+import com.task.task.databinding.ItemLayoutBinding
 
 
 class MoviesListingAdapter(private var actions: Action) : ListAdapter<ResultDomain, MovieItemViewHolder>(
@@ -15,7 +16,9 @@ class MoviesListingAdapter(private var actions: Action) : ListAdapter<ResultDoma
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieItemViewHolder {
         val viewHolder =
             LayoutInflater.from(parent.context).inflate(R.layout.item_layout, parent, false)
-        return MovieItemViewHolder(viewHolder)
+
+        val binding = ItemLayoutBinding.inflate(LayoutInflater.from(parent.context),parent,false)
+        return MovieItemViewHolder(binding)
     }
 
     override fun onBindViewHolder(holder: MovieItemViewHolder, position: Int) {
